+++ packages/block-editor/CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Experimental
 
--   Inserter: Behind the `gutenberg-media-inserter` experiment, render the Media tab as a single sidebar of drill-in panels, one per media source, in place of the category list and flyout panel ([#TBD](https://github.com/WordPress/gutenberg/pull/TBD)).
+-   Inserter: Behind the `gutenberg-media-inserter` experiment, render the Media tab as a single column of stacked, collapsible panels, one per media source, in place of the category list and flyout panel ([#TBD](https://github.com/WordPress/gutenberg/pull/TBD)).
 
 ### New Features